requirements.txt
streamlit==1.25.0
python-docx==0.8.11
pandas==2.2.2
openpyxl==3.1.2
matplotlib==3.7.2
